% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingFunctions.R
\name{logit}
\alias{logit}
\title{Logit of probability pr}
\usage{
logit(pr, use_ceiling_and_floor = TRUE)
}
\arguments{
\item{pr}{}
}
\description{
Adjusts for probability == 100% and probability == 0% by applying a ceiling and a floor
}
