% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ncv_auc}
\alias{ncv_auc}
\title{Run Nested Cross Validation on Training Data using AUC}
\usage{
ncv_auc(n, r, data, model, x)
}
\arguments{
\item{n}{Number of folds for the nested cross validation code.}

\item{r}{Number of repeats for the nested cross validation code.}

\item{data}{The dataframe with training data.}

\item{model}{Model used for nested cross validation code.}

\item{x}{Column name which represents the class you are trying to predict.}
}
\value{
Returns the AUC score of the nested cross validation code.
}
\description{
Run Nested Cross Validation on Training Data using AUC
}
\examples{
ncv_auc(2,1,train,"rf","Survived")
}

