% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nestedcv}
\alias{nestedcv}
\title{Run Nested Cross Validation on Training Data}
\usage{
nestedcv(n, r, data, model, x)
}
\arguments{
\item{n}{Number of folds for the nested cross validation code.}

\item{r}{Number of repeats for the nested cross validation code.}

\item{data}{The dataframe with training data.}

\item{model}{Model used for nested cross validation code}

\item{x}{Column name which represents the class you are trying to predict.}
}
\value{
Returns the accuracy score of the nested cross validation code.
}
\description{
Run Nested Cross Validation on Training Data
}
\examples{
nestedcv(2,1,train,"rf","Survived")
}

