% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloFunctions.R
\name{pm_eloOptimiseKfactor}
\alias{pm_eloOptimiseKfactor}
\title{pm_eloOptimiseKfactor}
\usage{
pm_eloOptimiseKfactor(eloSamp)
}
\arguments{
\item{eloSamp}{A dataset to optimise over}
}
\value{
A list of 3 parameters: Cfactor, Coffset, Cshape
}
\description{
Return a set of optimal Kfactors
}
