% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloFunctions.R
\name{pm_eloPrepDatabase}
\alias{pm_eloPrepDatabase}
\title{pm_eloPrepDatabase}
\usage{
pm_eloPrepDatabase(
  eloDB = NA,
  matchDB = NA,
  unseenDataPlayers,
  defaultScore = 1500,
  defaultMatches = 0
)
}
\arguments{
\item{eloDB}{a dictionary of player ELO scores, or NA to create a new one}

\item{matchDB}{a dictionary of matches a player has played, or NA}

\item{unseenDataPlayers}{list of players in the (as yet) unseen data}

\item{defaultScore}{the ELO for players without experience. Default of 1500}

\item{defaultMatches}{the number of matches for new players. Default of 0}
}
\value{
a list with two items: a dictionary for ELO scores, a dictionary for number of matches played
}
\description{
Prepare dictionaries for playerScore and playerMatches for unseen data (i.e. new players).
If the database does not exist (eloDB=NA) a new database will be generated.
}
\examples{


}
