% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloFunctions.R
\name{pm_eloReportElosAtDateFromDatabase}
\alias{pm_eloReportElosAtDateFromDatabase}
\title{pm_eloReportElosAtDateFromDatabase}
\usage{
pm_eloReportElosAtDateFromDatabase(
  eloDF,
  snapshotDate,
  numMonthsToBeIneligible = 3
)
}
\arguments{
\item{eloDF}{A dataframe containing elo_x data}

\item{snapshotDate}{A date for assessing the ELO scores}

\item{numMonthsToBeIneligible}{The number of months after which scores are ignored}
}
\value{
A dataframe with player_name and latest ELO rating at snapshotDate
}
\description{
Prepare dictionarys for playerScore and playerMatches for unseen data (i.e. new players).
If the database does not exist (eloDB=NA) a new database will be generated.
}
