% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloFunctions.R
\name{pm_eloReportTournamentStats}
\alias{pm_eloReportTournamentStats}
\title{pm_eloReportTournamentStats
Summarise a set of tournament results stored as one row per winner}
\usage{
pm_eloReportTournamentStats(df)
}
\arguments{
\item{df}{}
}
\value{
a tibble
}
\description{
pm_eloReportTournamentStats
Summarise a set of tournament results stored as one row per winner
}
