% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloFunctions.R
\name{pm_eloRunMultipleTimeSlices}
\alias{pm_eloRunMultipleTimeSlices}
\title{pm_eloRunMulitpleTimeSlices}
\usage{
pm_eloRunMultipleTimeSlices(eloDB, matchDB, eloDF, tennisElo = TRUE,
  Cfactor = 250, Coffset = 5, Cshape = 0.4)
}
\arguments{
\item{eloDB}{a hashmap of player ELO scores}

\item{matchDB}{a hashmap of matches a player has played}

\item{eloDF}{a dataframe of matches to process. Requires at least 'player_name', 'opponent_name', 'match_date' columns}
}
\value{
a list with three items: the updated eloDB, the updated matchDB and the input eloDF updated with ELO columns. New columns are prefixed with 'elo_'
}
\description{
Compute and update ELO scores for multiple time slices. Within each individual time slice all matches are assumed to be happening in parallel.
}
\examples{
tmpres = pm_eloPrepDatabase(unseenDataPlayers=c('bob','charlie'))
eloDB = tmpres$eloDB
matchDB = tmpres$matchDB
mysim = tibble::tribble(~player_name,~opponent_name,~match_date,~actualResult,
'bob','charlie',as.Date('2007-01-01'),1,
'bob','david',as.Date('2007-01-02'),0)
tmpres = pm_eloRunMultipleTimeSlices(eloDB,matchDB,mysim)

}
