% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloFunctions.R
\name{pm_eloRunTourneyELO}
\alias{pm_eloRunTourneyELO}
\title{pm_eloRunTourneyELO}
\usage{
pm_eloRunTourneyELO(tournamentSetup, keyCols = c("roundNum", "player_name",
  "opponent_name", "match_date", "Tournament"), simCols = c("predictions",
  "simWinner", "winnerName"), roundeloDB, roundmatchDB)
}
\arguments{
\item{tournamentSetup}{A dataframe strucutred to describe the tournament}

\item{keyCols}{Key columns for reporting results (DO NOT CHANGE FOR NOW)}

\item{simCols}{Key columsn for debugging results (DO NOT CHANGE FOR NOW)}

\item{eloDB}{a hashmap of player ELO scores, or NA to create a new one}

\item{matchDB}{a hashmap of matches a player has played, or NA}
}
\value{
A dataframe with winners (simWinner, winnerName) for each match in the Tournament
}
\description{
Run a simulated Tournament based on input ELO databases and a dataframe that describes the draw
}
\examples{

}
