% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloFunctions.R
\name{pm_eloShowWinProbForScoreDiff}
\alias{pm_eloShowWinProbForScoreDiff}
\title{pm_eloShowWinProbForScoreDiff}
\usage{
pm_eloShowWinProbForScoreDiff(scoreDiff)
}
\arguments{
\item{scoreDiff}{The difference in score}
}
\value{
The difference in odds due to the score difference
}
\description{
Return the probability of a win
}
