% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloFunctions.R
\name{pm_eloSimulateTournamentInParallel}
\alias{pm_eloSimulateTournamentInParallel}
\title{pm_eloSimulateTournamentInParallel}
\usage{
pm_eloSimulateTournamentInParallel(
  ds,
  runs = 10,
  eloDBForSim,
  matchDBForSim,
  keyCols = c("roundNum", "player_name", "opponent_name", "match_date", "Tournament"),
  simCols = c("predictions", "simWinner", "winnerName"),
  maxTimeSecs = 999999999999,
  tournamentName = "default",
  resetDatabase = FALSE,
  forceSingleThreaded = FALSE
)
}
\arguments{
\item{ds}{A dataframe structured to describe the tournament}

\item{runs}{Number of iterations (default = 10)}

\item{keyCols}{Key columns for reporting results (DO NOT CHANGE FOR NOW)}

\item{simCols}{Key columsn for debugging results (DO NOT CHANGE FOR NOW)}

\item{maxTimeSecs}{Maximum running time before it quites (defulat = a Very Big Number)}

\item{tournamentName}{The name of hte tournament _simulation_ (e.g. 'wimbledon 07')}

\item{resetDatabase}{Overwrite any existing predictions for tournamentName}

\item{forceSingleThreaded}{Force only a single thread to be used}

\item{eloDB}{a dictionary of player ELO scores, or NA to create a new one}

\item{matchDB}{a dictionary of matches a player has played, or NA}
}
\value{
A dataframe with winners (simWinner, winnerName) for each match in the Tournament
}
\description{
Sets up a parallel simulation of a tournament
}
\examples{

}
