% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubFunctions.R
\name{pm_gh_config}
\alias{pm_gh_config}
\title{pm_gh_config - configure github with username and email}
\usage{
pm_gh_config(github_username, github_email)
}
\arguments{
\item{github_username}{Your username for github}

\item{github_email}{Your github email address}
}
\description{
pm_gh_config - configure github with username and email
}
