% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubFunctions.R
\name{pm_gh_push_request}
\alias{pm_gh_push_request}
\title{Push your (previously committed) github changes to github}
\usage{
pm_gh_push_request()
}
\description{
Push your (previously committed) github changes to github
}
