% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingFunctions.R
\name{pm_modelling_autoclass_var}
\alias{pm_modelling_autoclass_var}
\title{pm_modelling_autoclass_var}
\usage{
pm_modelling_autoclass_var(dependent_var, independent_var, mycp = 0.005)
}
\arguments{
\item{mycp}{}
}
\description{
pm_modelling_autoclass_var
}
