% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingFunctions.R
\name{pm_modelling_calc_gini}
\alias{pm_modelling_calc_gini}
\title{pm_modelling_calc_gini}
\usage{
pm_modelling_calc_gini(df, predvar, truevar)
}
\arguments{
\item{truevar}{}
}
\description{
pm_modelling_calc_gini
}
