% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingFunctions.R
\name{pm_modelling_compare_player_opponent}
\alias{pm_modelling_compare_player_opponent}
\title{Ratios and differences of variable}
\usage{
pm_modelling_compare_player_opponent(df, varname)
}
\arguments{
\item{varname}{}
}
\description{
Ratios and differences of variable
}
