% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingFunctions.R
\name{pm_modelling_full_autoclass_var}
\alias{pm_modelling_full_autoclass_var}
\title{pm_modelling_full_autoclass_var}
\usage{
pm_modelling_full_autoclass_var(df, dependent_var, independent_var)
}
\arguments{
\item{independent_var}{}
}
\value{

}
\description{
pm_modelling_full_autoclass_var
}
