% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingFunctions.R
\name{pm_modelling_iv}
\alias{pm_modelling_iv}
\title{Calculate the information values for a variable in a dataframe}
\usage{
pm_modelling_iv(dv, iv)
}
\arguments{
\item{dependent_var}{}
}
\value{

}
\description{
Calculate the information values for a variable in a dataframe
}
