% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingFunctions.R
\name{pm_modelling_lifetime_performance}
\alias{pm_modelling_lifetime_performance}
\title{Workout lifetime win and loss rates}
\usage{
pm_modelling_lifetime_performance(df, group_variables = NA)
}
\arguments{
\item{group_variables}{vector of variables to calculate lifetime vars by (e.g. 'surface')}
}
\value{

}
\description{
Workout lifetime win and loss rates
}
