% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingFunctions.R
\name{pm_modelling_plot_variable}
\alias{pm_modelling_plot_variable}
\title{Plot the relationship between a variable and results}
\usage{
pm_modelling_plot_variable(df, var_name)
}
\arguments{
\item{var_name}{}
}
\description{
Plot the relationship between a variable and results
}
