% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingFunctions.R
\name{pm_modelling_score_dataframe}
\alias{pm_modelling_score_dataframe}
\title{pm_modelling_score_dataframe}
\usage{
pm_modelling_score_dataframe(df, scorecard, myone2one = 500,
  mypdo = 20, invert = FALSE)
}
\arguments{
\item{mypdo}{}
}
\description{
pm_modelling_score_dataframe
}
