% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingFunctions.R
\name{pm_modelling_tidify_player_opponent_dataframe}
\alias{pm_modelling_tidify_player_opponent_dataframe}
\title{Turn a dataframe with player_ and opponent_ data into a tidy dataframe}
\usage{
pm_modelling_tidify_player_opponent_dataframe(df, additional_variables = NA)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Both players and opponents become players in this new world!
}
