% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennisFunctions.R
\name{pm_tennis_eloify_dataset}
\alias{pm_tennis_eloify_dataset}
\title{Convert raw data into an Elo-friendly format}
\usage{
pm_tennis_eloify_dataset(my_raw_data)
}
\arguments{
\item{my_raw_data}{}
}
\description{
Convert raw data into an Elo-friendly format
}
