% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennisFunctions.R
\name{pm_tennis_fetchAllDatasets}
\alias{pm_tennis_fetchAllDatasets}
\title{Get all available historical data}
\usage{
pm_tennis_fetchAllDatasets(competition = "WTA")
}
\arguments{
\item{competition}{}
}
\description{
Get all available historical data
}
