% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennisFunctions.R
\name{pm_tennis_fetchDataset}
\alias{pm_tennis_fetchDataset}
\title{Fetch tennis results from www.tennis-data.co.uk}
\usage{
pm_tennis_fetchDataset(myyear, competition = "WTA")
}
\arguments{
\item{competition}{'WTA' by default}
}
\description{
Fetch tennis results from www.tennis-data.co.uk
}
\examples{
#'
}
