% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readfile}
\alias{readfile}
\title{Read dataframe file into R}
\usage{
readfile(file)
}
\arguments{
\item{file}{A csv file with training dataframe that you want to read.}
}
\value{
Returns training dataframe
}
\description{
Read dataframe file into R
}
\examples{
data<-readfile("train.csv")
}

